Gjenbruk av steg 2 - Bosituasjon
