+++ src/barnetilsyn/steg/2-bosituasjon/Bosituasjon.tsx
@@ -1,28 +1,11 @@
-import React, { FC, useEffect, useState } from 'react';
-import { FormattedHTMLMessage, useIntl } from 'react-intl';
-import LocaleTekst from '../../../language/LocaleTekst';
-import MultiSvarSpørsmål from '../../../components/spørsmål/MultiSvarSpørsmål';
-import SeksjonGruppe from '../../../components/gruppe/SeksjonGruppe';
+import React, { FC } from 'react';
+import { useIntl } from 'react-intl';
 import Side from '../../side/Side';
-import SøkerSkalFlytteSammenEllerFåSamboer from './SøkerSkalFlytteSammenEllerFåSamboer';
-import { delerSøkerBoligMedAndreVoksne } from './BosituasjonConfig';
-import { erValgtSvarLiktSomSvar } from '../../../utils/spørsmålogsvar';
-import { hentTekst } from '../../../utils/søknad';
-import { ISpørsmål, ISvar } from '../../../models/spørsmålogsvar';
-import {
-  ESøkerDelerBolig,
-  IBosituasjon,
-} from '../../../models/steg/bosituasjon';
-import { useHistory, useLocation } from 'react-router-dom';
-import { Hovedknapp } from 'nav-frontend-knapper';
-import FeltGruppe from '../../../components/gruppe/FeltGruppe';
+import { IBosituasjon } from '../../../models/steg/bosituasjon';
+import { useLocation } from 'react-router-dom';
 import { erFerdigUtfylt } from '../../../helpers/steg/bosituasjon';
-
-import EkteskapsliknendeForhold from './EkteskapsliknendeForhold';
-import OmTidligereSamboer from './OmTidligereSamboer';
-import AlertStripeDokumentasjon from '../../../components/AlertstripeDokumentasjon';
-import AlertStripe from 'nav-frontend-alertstriper';
 import { useBarnetilsynSøknad } from '../../BarnetilsynContext';
+import BosituasjonSpørsmål from '../../../søknad/steg/2-bosituasjon/BosituasjonSpørsmål';
 
 const Bosituasjon: FC = () => {
   const intl = useIntl();
@@ -32,55 +15,18 @@ const Bosituasjon: FC = () => {
     settDokumentasjonsbehov,
     mellomlagreOvergangsstønad,
   } = useBarnetilsynSøknad();
-  const history = useHistory();
+  const bosituasjon = søknad.bosituasjon;
   const location = useLocation();
   const kommerFraOppsummering = location.state?.kommerFraOppsummering;
-  const hovedSpørsmål: ISpørsmål = delerSøkerBoligMedAndreVoksne;
-  const [bosituasjon, settBosituasjon] = useState<IBosituasjon>(
-    søknad.bosituasjon
-  );
 
-  useEffect(() => {
-    settSøknad({ ...søknad, bosituasjon: bosituasjon });
-    // eslint-disable-next-line
-  }, [bosituasjon]);
-
-  const settBosituasjonFelt = (spørsmål: ISpørsmål, svar: ISvar) => {
-    const svarTekst: string = hentTekst(svar.svar_tekstid, intl);
-    const spørsmålTekst: string = hentTekst(spørsmål.tekstid, intl);
-
-    const nyBosituasjon = {
-      delerBoligMedAndreVoksne: {
-        spørsmålid: spørsmål.søknadid,
-        svarid: svar.id,
-        label: spørsmålTekst,
-        verdi: svarTekst,
-      },
-    };
-
-    settBosituasjon(nyBosituasjon);
-    settDokumentasjonsbehov(spørsmål, svar);
+  const settBosituasjon = (bosituasjon: IBosituasjon) => {
+    settSøknad((prevSoknad) => {
+      return {
+        ...prevSoknad,
+        bosituasjon: bosituasjon,
+      };
+    });
   };
-
-  const valgtSvar:
-    | ISvar
-    | undefined = hovedSpørsmål.svaralternativer.find((svar) =>
-    erValgtSvarLiktSomSvar(
-      bosituasjon.delerBoligMedAndreVoksne.verdi,
-      svar.svar_tekstid,
-      intl
-    )
-  );
-
-  const harSøkerEkteskapsliknendeForhold =
-    bosituasjon.delerBoligMedAndreVoksne.svarid ===
-    ESøkerDelerBolig.harEkteskapsliknendeForhold;
-
-  const planerOmÅFlytteSammenEllerFåSamboer =
-    bosituasjon.delerBoligMedAndreVoksne.svarid ===
-      ESøkerDelerBolig.borAleneMedBarnEllerGravid ||
-    bosituasjon.delerBoligMedAndreVoksne.svarid ===
-      ESøkerDelerBolig.delerBoligMedAndreVoksne;
 
   return (
     <Side
@@ -89,68 +35,11 @@ const Bosituasjon: FC = () => {
       erSpørsmålBesvart={erFerdigUtfylt(bosituasjon)}
       mellomlagreOvergangsstønad={mellomlagreOvergangsstønad}
     >
-      <SeksjonGruppe>
-        <MultiSvarSpørsmål
-          key={hovedSpørsmål.søknadid}
-          spørsmål={hovedSpørsmål}
-          valgtSvar={bosituasjon.delerBoligMedAndreVoksne.verdi}
-          settSpørsmålOgSvar={settBosituasjonFelt}
-        />
-        {valgtSvar && valgtSvar.alert_tekstid && (
-          <FeltGruppe>
-            {bosituasjon.delerBoligMedAndreVoksne.svarid ===
-            ESøkerDelerBolig.tidligereSamboerFortsattRegistrertPåAdresse ? (
-              <AlertStripeDokumentasjon>
-                <FormattedHTMLMessage id={valgtSvar.alert_tekstid} />
-              </AlertStripeDokumentasjon>
-            ) : (
-              <AlertStripe type={'advarsel'} form={'inline'}>
-                <LocaleTekst tekst={valgtSvar.alert_tekstid} />
-              </AlertStripe>
-            )}
-          </FeltGruppe>
-        )}
-      </SeksjonGruppe>
-
-      {planerOmÅFlytteSammenEllerFåSamboer && (
-        <SeksjonGruppe>
-          <SøkerSkalFlytteSammenEllerFåSamboer
-            settBosituasjon={settBosituasjon}
-            bosituasjon={bosituasjon}
-          />
-        </SeksjonGruppe>
-      )}
-
-      {harSøkerEkteskapsliknendeForhold && (
-        <EkteskapsliknendeForhold
-          settBosituasjon={settBosituasjon}
-          bosituasjon={bosituasjon}
-        />
-      )}
-
-      {bosituasjon.delerBoligMedAndreVoksne.svarid ===
-        ESøkerDelerBolig.tidligereSamboerFortsattRegistrertPåAdresse && (
-        <SeksjonGruppe>
-          <OmTidligereSamboer
-            bosituasjon={bosituasjon}
-            settBosituasjon={settBosituasjon}
-          />
-        </SeksjonGruppe>
-      )}
-      {kommerFraOppsummering && erFerdigUtfylt(bosituasjon) ? (
-        <div className={'side'}>
-          <Hovedknapp
-            className="tilbake-til-oppsummering"
-            onClick={() =>
-              history.push({
-                pathname: '/oppsummering',
-              })
-            }
-          >
-            {hentTekst('oppsummering.tilbake', intl)}
-          </Hovedknapp>
-        </div>
-      ) : null}
+      <BosituasjonSpørsmål
+        bosituasjon={bosituasjon}
+        settBosituasjon={settBosituasjon}
+        settDokumentasjonsbehov={settDokumentasjonsbehov}
+      />
     </Side>
   );
 };
