+++ src/søknad/steg/2-bosituasjon/SøkerSkalFlytteSammenEllerFåSamboer.tsx
@@ -12,20 +12,24 @@ import { useIntl } from 'react-intl';
 import { hentBooleanFraValgtSvar } from '../../../utils/spørsmålogsvar';
 import { hentTekst } from '../../../utils/søknad';
 import { ISvar } from '../../../models/spørsmålogsvar';
-import { useSøknad } from '../../../context/SøknadContext';
 import { datoTilStreng } from '../../../utils/dato';
 
 interface Props {
   settBosituasjon: (bosituasjon: IBosituasjon) => void;
   bosituasjon: IBosituasjon;
+  settDokumentasjonsbehov: (
+    spørsmål: ISpørsmål,
+    valgtSvar: ISvar,
+    erHuketAv?: boolean
+  ) => void;
 }
 
 const SøkerSkalFlytteSammenEllerFåSamboer: FC<Props> = ({
   settBosituasjon,
   bosituasjon,
+  settDokumentasjonsbehov,
 }) => {
   const intl = useIntl();
-  const { settDokumentasjonsbehov } = useSøknad();
 
   const {
     delerBoligMedAndreVoksne,
